moving things around.
While when I started this project,
it's meant to be experimental so I don't bother
with throwing all sorts of random stuff at the problem.
but now that I think about it, this is a perfect oppotunity
to grow this into something better - kantour-core can be a
core library that other things can be built on top of it and
kantour CLI can still keep it experimental state.